Complete registration functionality

+++ base/views.py
@@ -2,14 +2,21 @@ from django.shortcuts import render, redirect
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
 
 
 def login_page(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        messages.info(request, 'you are already logged in!!')
+        return redirect('home')
+    
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -24,7 +31,24 @@ def login_page(request):
         else:
             messages.error(request, 'Invalid credentials')
 
-    context = {}
+    context = { 'page': page }
+    return render(request, 'base/login_register.html', context)
+
+
+def register_user(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error was occurred during the registration')
+
+    context = { 'form': form }
     return render(request, 'base/login_register.html', context)
 
 
@@ -54,6 +78,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def new_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -67,9 +92,14 @@ def new_room(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(pk=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        messages.error(request, 'You are not allowed to edit this room!!')
+        return redirect('home')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -81,8 +111,14 @@ def update_room(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(pk=pk)
+
+    if request.user != room.host:
+        messages.error(request, 'You are not allowed to edit this room!!')
+        return redirect('home')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
